Rename typo in folder
